LDUR and STUR work (haven't tested edge cases yet)

+++ src/instructions/d_type_funcs.c
@@ -5,7 +5,8 @@ void instruction_ldur(emulator_t* emulator, uint8_t* instruction) {
   uint16_t address;
   uint8_t op2, rn, rt;
   get_d_format_params(instruction, &address, &op2, &rn, &rt);
-  int8_t result = get_data(emulator, address, rn);
+  int64_t result = get_data(emulator, address, rn);
+
   set_reg(emulator, rt, result, 0);
   printf("LDUR called: X%d = X%d[%d] with op:%d.\n", rt, rn, address, op2);
 }
